requirements.txt
attrs==17.3.0
beautifulsoup4==4.6.0
boto3==1.5.14
botocore==1.8.28
bs4==0.0.1
dj-database-url==0.4.2
Django==2.0
django-cors-headers==2.1.0
django-extra-fields==0.9
django-storages==1.6.5
djangorestframework==3.7.3
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
nltk==3.2.5
numpy==1.14.1
olefile==0.44
Pillow==4.3.0
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pytest==3.3.1
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.12
six==1.11.0
whitenoise==3.3.1
